+++ backend/annotation/serializers_test.py
@@ -5,46 +5,13 @@ from django.utils import timezone
 from django.contrib.auth.models import Permission
 from rest_framework.test import APIRequestFactory
 
-from annotation.models import KnowledgeBaseAnnotation, Label, LabelAnnotation
 from annotation.serializers import (
     KnowledgeBaseAnnotationSerializer,
     LabelSerializer,
     LabelAnnotationSerializer,
-    AnnotationSerializer,
     SaveLabelsInputSerializer,
 )
-
-
-@pytest.fixture
-def kb_annotation(db, sample_problem, annotation_session):
-    return KnowledgeBaseAnnotation.objects.create(
-        problem=sample_problem,
-        entity1="e1",
-        entity2="e2",
-        relationship=KnowledgeBaseAnnotation.Relationship.EQUAL,
-        session=annotation_session,
-        created_at="2024-01-01T00:00:00Z",
-        created_by=annotation_session.user,
-    )
-
-
-@pytest.fixture
-def sample_label(db):
-    """Create a sample label for testing."""
-    return Label.objects.create(
-        text="Ambiguous", description="This problem contains ambiguous language."
-    )
-
-
-@pytest.fixture
-def label_annotation(db, sample_problem, annotation_session, sample_label):
-    """Create a label annotation for testing."""
-    return LabelAnnotation.objects.create(
-        problem=sample_problem,
-        label=sample_label,
-        session=annotation_session,
-        created_by=annotation_session.user,
-    )
+from problem.serializers import ProblemSerializer
 
 
 @pytest.mark.django_db
@@ -118,8 +85,7 @@ def test_kb_annotation_update(kb_annotation):
 def test_kb_annotation_get_removable_with_permission(kb_annotation, annotator):
     """Test that get_removable returns True when user has permission."""
     permission = Permission.objects.get(
-        codename="delete_knowledgebaseannotation",
-        content_type__app_label="annotation"
+        codename="delete_knowledgebaseannotation", content_type__app_label="annotation"
     )
     annotator.user_permissions.add(permission)
     annotator.refresh_from_db()
@@ -214,26 +180,6 @@ def test_label_annotation_not_removable_without_permission(label_annotation, vis
 
 
 @pytest.mark.django_db
-def test_annotation_serializer_with_kb_and_labels(
-    sample_problem, annotation_session, kb_annotation, label_annotation
-):
-    """Test AnnotationSerializer returns both KB and label annotations."""
-    serializer = AnnotationSerializer(
-        {},
-        context={
-            "problem": sample_problem,
-            "user": annotation_session.user,
-        },
-    )
-    data: dict[str, Any] = serializer.data  # type: ignore
-
-    assert "kbAnnotations" in data
-    assert "labelAnnotations" in data
-    assert len(data["kbAnnotations"]) > 0
-    assert len(data["labelAnnotations"]) > 0
-
-
-@pytest.mark.django_db
 def test_annotation_serializer_excludes_removed_annotations(
     sample_problem, annotation_session, kb_annotation
 ):
@@ -243,13 +189,7 @@ def test_annotation_serializer_excludes_removed_annotations(
     kb_annotation.removed_by = annotation_session.user
     kb_annotation.save()
 
-    serializer = AnnotationSerializer(
-        {},
-        context={
-            "problem": sample_problem,
-            "user": annotation_session.user,
-        },
-    )
+    serializer = ProblemSerializer(sample_problem)
     data: dict[str, Any] = serializer.data  # type: ignore
 
     assert len(data["kbAnnotations"]) == 0
